+++ src/apps/Chat/containers/Right/Styles.ts
@@ -1,17 +1,10 @@
-import {Form as UForm} from 'formik';
 import styled from 'styled-components';
 
 import {Status} from 'apps/Chat/components/Avatar/Styles';
-import UButton from 'apps/Chat/components/Button';
-import {InlineInput} from 'apps/Chat/components/FormElements';
 import {colors, fonts} from 'apps/Chat/styles';
 
 export const BottomMessage = styled.div`
   height: 8px;
-`;
-
-export const Button = styled(UButton)`
-  display: none;
 `;
 
 export const Container = styled.div`
@@ -21,21 +14,11 @@ export const Container = styled.div`
   flex-direction: column;
 `;
 
-export const ContentInput = styled(InlineInput)`
-  flex: 1;
-`;
-
 export const EmptyState = styled.div`
   align-items: center;
   background: ${colors.rightBackground};
   display: flex;
   justify-content: center;
-`;
-
-export const Form = styled(UForm)`
-  align-items: center;
-  display: flex;
-  padding: 0 16px 16px;
 `;
 
 export const Messages = styled.div`
